Support OCI object storage for logging and reports.
